refactor : api연동 컴포넌트 분리
